fix : CPUView

+++ src/backends/cpu/CPUView.hpp
@@ -29,8 +29,8 @@ private:
 class CPUViewCreator : public CPUBackend::Creator {
 public:
     virtual Op *create(OpParam op_param, Backend *bn, string name) const {
-        vector<int> dims = {op_param["dim0"], op_param["dim1"], op_param["dim2"], op_param["dim3"]};
-        vector<int> data_dims = {op_param["data_dim0"], op_param["data_dim1"], op_param["data_dim2"], op_param["data_dim3"]};
+        vector<int> dims = {(int)op_param["dim0"], (int)op_param["dim1"], (int)op_param["dim2"], (int)op_param["dim3"]};
+        vector<int> data_dims = {(int)op_param["data_dim0"], (int)op_param["data_dim1"], (int)op_param["data_dim2"], (int)op_param["data_dim3"]};
         return new CPUView(bn, name, dims, data_dims, false);
     }
 };
